requirements.txt

bcrypt==4.3.0
click==8.2.1
email_validator==2.2.0
fastapi==0.115.12
passlib==1.7.4
pydantic==2.11.7
pymongo==4.13.1
python-dotenv==1.1.0
python-jose==3.5.0
uvicorn==0.34.3
